NANO-458: Allow user to have speculative reads
Add a command line paramter for the flatten-cfg pass that allows the
user to tell the pass to move packet and map reads without making them
conditional.  As a result, they will execute speculatively.  That is
fine from the packet / map perspective, but can be problematic, if the
application reuses the read buffers and uses across multiple reads.
Then a read that would have been behind a conditional branch will now
awlways execute and overwrite data in the read buffer.