NANO-459: Improve error & debug msg, comment
Improve the debug printout, error messages, and comments in the
code that handles moving / flattening calls.  We need to either
be able to prove that a call is side-effect free, be able to
predicate it, or be able to prove that it was always executed in the
original program.  Make these cases a bit more visible for easier
debug and understanding of the error messages.