NANO-460: Replacing several abort() with exit(1)
To make error reporting a bit more useful, remove calls to abort()
which will create a compiler stack trace when that is not useful.
The original idea was to separate user code errors vs inconsistencies
in the compiler, but there remains a number of in-between cases where
the compiler might need extension.  I have followed the thinking of
"is a stack trace here helpful or not" when changing the calls.