Testing: Use relative paths for running tests
Change the LIT script and run_*_test helpers to handle relative paths
for the tests.  Also change the logic which computes the different paths
for a given test source file.  As a result, the printed command lines
for running tests is much shorter due to using relative paths.

The new parsing logic that computes directories from the test name now
computes the following:
- test source location and golden data that does not need building ->
  from the provided test source name parameter
- built test executables and golden data that has been built -> from a
  path relative to the run_*_test script location